superset/docker/requirements.txt
flask-oidc==1.3.0
itsdangerous==2.0.1
flask_openid
authlib
pymysql 
psycopg2-binary
clickhouse-connect 
sqlalchemy-vertica-python